server/requirements.txt
livekit-agents
livekit
numpy
soundfile
python-dotenv
transformers
sentencepiece
torch
faster-whisper
webrtcvad
websockets
requests
google-cloud-texttospeech
